Included AudioType.network networkHeaders
The reason for omitting AudioType.network in centralized networkSettings.defaultHeaders is unknown. Hence, I added the same to the existing condition.
I has a requirement to send headers to the server during audio fetching via AudioType.network. Found out, AudioType.network was missing and now the headers are working fine for me.